requirements.txt
torch==2.2.2
torchvision==0.17.2
mlflow==3.10.1
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.8.4
tqdm==4.66.2
boto3==1.34.84
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
evidently==0.4.22
pydantic==2.7.1
